As Finished as Possible
Some components are missing from this project such as the lighting
because of the changes from Unity 4.6 to Unity 5. In the textbook it
told me to refer back to the website where their would be different
downloadable content for the project but nothing was added to the site.

Another thing missing is the crosshairs in the top left screen, I had to
fiddle around with multiple camera to make the crosshair on the FPCamera
work, but could not get it to work for the other camera.